Place statistic card into a separate component
